Extracted Session logic from middeleware into Session controller
