+++ apps/augury-backend/src/controllers/auth/SessionController.ts
@@ -1,32 +1,54 @@
 import mongoose, { Types } from 'mongoose';
 import { AxiosError } from 'axios';
-import Session from '../../config/interfaces/Session';
-import ApiError from '../../errors/ApiError';
+import { verifyJwt } from '../../config/utils/jwt';
 import SessionModel from '../../models/auth/SessionModel';
+import Session from '../../config/interfaces/Session';
+import StatusCode from '../../config/enums/StatusCode';
+import Severity from '../../config/enums/Severity';
+import ApiError from '../../errors/ApiError';
 
 export async function getSession(id: Types.ObjectId | string, token: string) {
   const _id = new mongoose.Types.ObjectId(id);
-  let response;
   const session: Session = {
     userId: _id,
     token: token,
   };
   try {
-    response = await SessionModel.updateSession(session);
+    const response = await SessionModel.updateSession(session);
+    return response;
   } catch (error: unknown) {
     if (error instanceof ApiError) {
-      response = await SessionModel.createSession(session);
+      const response = await SessionModel.createSession(session);
+      return response;
     } else if (error instanceof AxiosError) {
       throw new Error(error.message);
     } else {
       throw new Error(`Unknown error occurred! ${JSON.stringify(error)}`);
     }
   }
-
-  return response;
 }
 
-export async function getSessionByToken(token: string) {
-  const session = await SessionModel.getSessionByToken(token);
-  return session;
+export async function getSessionByToken(accessToken: string) {
+  try {
+    // Verify and decode the JWT
+    const verificationResult = verifyJwt(accessToken);
+    if (
+      typeof verificationResult.decoded === 'string' ||
+      !verificationResult?.decoded?.session
+    ) {
+      throw new Error();
+    }
+    const token = verificationResult.decoded.session;
+    const session = await SessionModel.getSessionByToken(token);
+    if (!session) {
+      throw new Error(); // Invalid Session
+    }
+    return session;
+  } catch (error: unknown) {
+    throw new ApiError(
+      'Invalid session token or incorrect token structure.',
+      StatusCode.FORBIDDEN,
+      Severity.MED
+    );
+  }
 }
