+++ apps/augury-backend/src/middlewares/VerifyAccessTokenHandler.ts
@@ -1,11 +1,9 @@
 import { Request, Response, NextFunction } from 'express';
-import { verifyJwt } from '../config/utils/jwt';
 import ApiError from '../errors/ApiError';
 import ClientError from '../errors/ClientError';
 import StatusCode from '../config/enums/StatusCode';
 import Severity from '../config/enums/Severity';
-import Session from '../config/interfaces/Session';
-import SessionModel from '../models/auth/SessionModel';
+import { getSessionByToken } from '../controllers/auth/SessionController';
 
 export async function verifyAccessToken(
   req: Request,
@@ -18,19 +16,7 @@ export async function verifyAccessToken(
   }
 
   try {
-    // Verify and decode the JWT
-    const verificationResult = verifyJwt(accessToken);
-    if (
-      typeof verificationResult.decoded === 'string' ||
-      !verificationResult?.decoded?.session
-    ) {
-      throw new Error();
-    }
-    const token = verificationResult.decoded.session;
-    const session: Session = await SessionModel.getSessionByToken(token);
-    if (!session) {
-      throw new Error(); // Invalid Session
-    }
+    getSessionByToken(accessToken);
     next();
   } catch (error) {
     throw new ApiError(
